+++ packages/salesforcedx-sobjects-faux-generator/src/generator/types.ts
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2021, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export interface FieldDeclaration {
+  modifier: string;
+  type: string;
+  name: string;
+  comment?: string;
+}
+
+export interface SObjectDefinition {
+  name: string;
+  fields: FieldDeclaration[];
+}
